% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze counterfactual explanations using a Shiny app}
\usage{
analyze(counterfactual_explanations, digits = 2)
}
\arguments{
\item{counterfactual_explanations}{An object of class \code{counterfactual_explanations} containing the original observation, generated counterfactuals, and objective values.}

\item{digits}{Number of digits to round the numerical values to.}
}
\description{
Analyze counterfactual explanations using a Shiny app
}
