% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{analyze_clustering}
\alias{analyze_clustering}
\title{Analyze clustering of time-dependent survival predictions}
\usage{
analyze_clustering(
  explainer,
  clustering,
  max_k = 10,
  min_obs = 10,
  p = 0,
  q = 0
)
}
\arguments{
\item{explainer}{A \code{survex::survival_explainer} object - model preprocessed by the \code{survex::explain()} function.}

\item{clustering}{A hierarchical clustering object.}

\item{max_k}{An integer value indicating the maximum number of clusters to be considered.}

\item{min_obs}{An integer value indicating the minimum number of observations in each cluster.}

\item{p}{A numeric value being a weight used for the power of Kaplan-Meier estimator values in the Fleming-Harrington test.}

\item{q}{A numeric value being a weight used for the power of 1 - Kaplan-Meier estimator values in the Fleming-Harrington test.}
}
\description{
Analyze clustering of time-dependent survival predictions for different numbers of clusters using the Fleming-Harrington test and plots showing the p-values of the test and the dendrogram.
}
