% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_distance.R
\name{distance_from_target_prediction_band}
\alias{distance_from_target_prediction_band}
\title{Calculate distance from target prediction band}
\usage{
distance_from_target_prediction_band(
  sf,
  prediction_band,
  times,
  weights = rep(1/length(times), length(times))
)
}
\arguments{
\item{sf}{Time-dependent survival predictions for which the distance from the target prediction band is calculated.}

\item{prediction_band}{An object of class \code{target_prediction_band} containing the target prediction band for the counterfactuals.}

\item{times}{A numeric vector of time points at which the distance is calculated.}

\item{weights}{A numeric vector of weights assigned to each time point.}
}
\description{
Calculate distance from target prediction band
}
