% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_clusters}
\alias{get_clusters}
\title{Get clusters}
\usage{
get_clusters(dendrogram, k)
}
\arguments{
\item{dendrogram}{A hierarchical clustering object.}

\item{k}{An integer value indicating the number of clusters.}
}
\description{
Get clusters from a hierarchical clustering object.
}
