% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_hierarchical_clustering}
\alias{get_hierarchical_clustering}
\title{Get hierarchical clustering of time-dependent survival predictions}
\usage{
get_hierarchical_clustering(
  explainer,
  weights,
  output_type = "survival",
  linkage_method = "ward.D2"
)
}
\arguments{
\item{explainer}{A \code{survex::survival_explainer} object - model preprocessed by the \code{survex::explain()} function.}

\item{weights}{A numeric vector of weights assigned to each time point in explainer$times.}

\item{output_type}{A character value indicating the type of survival predictions to be used in the clustering. It can be either "survival" for survival function or "chf" for cumulative hazard function.}

\item{linkage_method}{A character value indicating the method used to calculate the distance between clusters. It can be one of "ward.D2", "single", "complete", "average", "mcquitty", "median" or "centroid".}
}
\description{
Get hierarchical clustering of time-dependent survival predictions
}
