% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_band.R
\name{get_prediction_band}
\alias{get_prediction_band}
\title{Get prediction band}
\usage{
get_prediction_band(
  x,
  clusters,
  cluster_id,
  lambda = 0.05,
  output_type = "survival"
)
}
\arguments{
\item{x}{A numeric matrix of predictions or a \code{survex::survival_explainer} object containing the model and data.}

\item{clusters}{A numeric vector of cluster assignments.}

\item{cluster_id}{An integer value indicating the cluster for which the prediction band is calculated.}

\item{lambda}{A numeric value indicating the positional quantile of the predictions used to calculate the prediction band bounds.}

\item{output_type}{A character value indicating the type of predictions. Either "survival" or "chf".}
}
\description{
Get prediction band for a specific cluster.
}
