% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovalev_method.R
\name{kovalev_method}
\alias{kovalev_method}
\title{Kovalev method}
\usage{
kovalev_method(
  explainer,
  new_observation,
  r,
  num_iter = 200,
  num_particles = 1000,
  verbose = 10,
  C = 1e+06,
  w = 0.729,
  c1 = 1.4945,
  c2 = 1.4945,
  tol_iter = 10,
  seed = NULL
)
}
\arguments{
\item{explainer}{A \code{survex::survival_explainer} object - model preprocessed by the \code{survex::explain()} function.}

\item{new_observation}{A data frame with a single row containing the observation for which counterfactuals are to be generated.}

\item{r}{The target difference between the mean time to survival of the counterfactual and the original observation.}

\item{num_iter}{Number of iterations in the PSO algorithm.}

\item{num_particles}{Number of particles in the PSO algorithm.}

\item{verbose}{Integer value indicating how often the loss should be printed.}

\item{C}{Penalty parameter for the loss function (used to enforce the counterfactuality).}

\item{w}{Inertia weight in PSO algorithm.}

\item{c1}{Cognitive parameter in PSO algorithm.}

\item{c2}{Social parameter in PSO algorithm.}

\item{tol_iter}{Integer value indicating the number of iterations without progress after which the search should be stopped.}

\item{seed}{An integer value to set the seed for reproducibility.}
}
\description{
Kovalev method is a model-agnostic counterfactual explanation approach for survival models. It is based on the Particle Swarm Optimization algorithm.
}
