% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_changes_frequency}
\alias{plot_changes_frequency}
\title{Plot frequency of variable changes in counterfactual explanations}
\usage{
plot_changes_frequency(
  counterfactual_explanations,
  filtered_examples = NULL,
  variables = NULL,
  max_variables = Inf
)
}
\arguments{
\item{counterfactual_explanations}{a \code{counterfactual_explanations} object with counterfactual examples, objective values and original observation}

\item{filtered_examples}{a data frame with counterfactual examples. If NULL, all counterfactual examples from \code{counterfactual_explanations} will be used}

\item{variables}{a numeric vector of column indices representing variables to plot. If NULL, all variables will be on the plot}

\item{max_variables}{a numeric value indicating the maximum number of variables to plot}
}
\description{
Plot frequency of variable changes in counterfactual explanations
}
