% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_parallel_coordinates}
\alias{plot_parallel_coordinates}
\title{Plot parrallel coordinates plot for counterfactual explanations}
\usage{
plot_parallel_coordinates(
  counterfactual_explanations,
  filtered_examples = NULL,
  variables = NULL
)
}
\arguments{
\item{counterfactual_explanations}{a \code{counterfactual_explanations} object with counterfactual explanations}

\item{filtered_examples}{a data frame with counterfactual examples. If NULL, all counterfactual examples from \code{counterfactual_explanations} will be used}

\item{variables}{a numeric vector of column indices representing variables to plot. If NULL, all variables will be on the plot}
}
\description{
Plot parrallel coordinates plot for counterfactual explanations
}
