% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_band.R
\name{plot_prediction_bands}
\alias{plot_prediction_bands}
\title{Plot prediction bands}
\usage{
plot_prediction_bands(
  explainer,
  clusters,
  lambda = 0.05,
  preds_to_plot = NULL,
  alpha = 0.2,
  output_type = "survival"
)
}
\arguments{
\item{explainer}{A \code{survex::survival_explainer} object - model preprocessed by the \code{survex::explain()} function.}

\item{clusters}{A numeric vector of cluster assignments.}

\item{lambda}{A numeric value indicating the positional quantile of the predictions used to calculate the prediction band bounds.}

\item{preds_to_plot}{A numeric matrix of single predictions to be plotted.}

\item{alpha}{A numeric value indicating the transparency of the prediction bands.}

\item{output_type}{A character value indicating the type of predictions. Either "survival" or "chf".}
}
\description{
Plot prediction bands
}
