% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot time-dependent predictions}
\usage{
plot_predictions(
  preds,
  times,
  preds_labels = NULL,
  alpha = 0.3,
  linewidth = 0.3
)
}
\arguments{
\item{preds}{a matrix of predictions}

\item{times}{a vector of time points at which the predictions were made}

\item{preds_labels}{a vector of labels for the predictions}

\item{alpha}{transparency of the lines}

\item{linewidth}{width of the lines}
}
\description{
Plot time-dependent predictions
}
