% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{plot_survival_weights}
\alias{plot_survival_weights}
\title{Plot Fleming-Harrington survival weights}
\usage{
plot_survival_weights(explainer, times, p = 0, q = 0, add_mean = FALSE)
}
\arguments{
\item{explainer}{A \code{survex::survival_explainer} object - model preprocessed by the \code{survex::explain()} function.}

\item{times}{A numeric vector of time points at which the weights are calculated.}

\item{p}{A numeric vector of weights used for the power of Kaplan-Meier estimator values.}

\item{q}{A numeric vector of weights used for the power of 1 - Kaplan-Meier estimator values.}

\item{add_mean}{A logical value indicating whether to add the mean of the weights to the weights.}
}
\description{
Plot Fleming-Harrington survival weights
}
