% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_pt.R
\name{st_pt}
\alias{st_pt}
\title{ST-PT: Survival Trees - Path Tweaking}
\usage{
st_pt(
  explainer,
  new_observation,
  target_prediction_band = NULL,
  weights = survival_weights(explainer, p = 0, q = 0),
  times = explainer$times,
  background_data = NULL,
  paths_per_tree = 20L,
  paths_per_counterfactual = 1L,
  max_tries = 200,
  step = 1,
  max_counterfactuals = NULL,
  fixed_variables_indices = NULL,
  plausible_values = NULL,
  verbose = FALSE,
  seed = NULL
)
}
\arguments{
\item{explainer}{A \code{survex::survival_explainer} object - model preprocessed by the \code{survex::explain()} function.}

\item{new_observation}{A data frame with a single row containing the observation for which counterfactuals are to be generated.}

\item{target_prediction_band}{An object of class \code{target_prediction_band} containing the target prediction band for the counterfactuals.}

\item{weights}{A numeric vector with non-negative values and sum(weights) > 0. Weights for each time point. Default is \code{survex::survival_weights(explainer, p=0, q=0)}, which assigns equal weights to all time points.}

\item{times}{A numeric vector with time points for which the predictions are to be calculated.}

\item{background_data}{A data frame with the background data used to train the model. If not provided, the data from the explainer object is used.}

\item{paths_per_tree}{Number of paths to be extracted from each tree. Default is 20.}

\item{paths_per_counterfactual}{Number of paths to be combined to create a single counterfactual. Default is 1, which means that a process is deterministic. If set to a value greater than 1, the paths are randomly combined to create counterfactuals.}

\item{max_tries}{Maximum number of tries to generate a counterfactual, used only when \code{paths_per_counterfactual} > 1. Default is 200.}

\item{step}{Step size for tweaking the values of features. Default is 1.}

\item{max_counterfactuals}{Maximum number of counterfactuals to be returned. Default is 40 or \code{num_trees * paths_per_tree}, whichever is smaller.}

\item{fixed_variables_indices}{A vector of positive integers indicating the indices of variables that should remain fixed during the counterfactual generation process. Default is \code{NULL}.}

\item{plausible_values}{A list with names corresponding to column names of \code{new_observation}. Each element of the list is a vector of plausible values for the corresponding feature. Default is \code{NULL}, which means that all unique values from the background data are considered plausible.}

\item{verbose}{Logical indicating whether to print progress messages. Default is \code{FALSE}.}

\item{seed}{An integer to set the random seed for reproducibility. Default is \code{NULL}.}
}
\description{
ST-PT is a time-dependent model-specific counterfactual explanation method tailored for tree-based ensembles, particularly random survival forests. It generates counterfactuals based on extraction of counterfactual decision paths from trees
}
