% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_moc.R
\name{surv_moc}
\alias{surv_moc}
\title{SurvMOC: Survival Multi-Objective Counterfactuals}
\usage{
surv_moc(
  explainer,
  new_observation,
  target_prediction_band = NULL,
  weights = survival_weights(explainer, p = 0, q = 0),
  times = explainer$times,
  background_data = NULL,
  target_change = NULL,
  fixed_variables_indices = NULL,
  numerical_variables_indices = NULL,
  categorical_variables_indices = NULL,
  integer_variables_indices = NULL,
  binary_variables_indices = NULL,
  plausible_data_range = NULL,
  plausible_categorical_values = NULL,
  initialization_type = "ice",
  data_distance_p = 1,
  plausibility_k_neighbors = 5,
  sbx_eta = 5,
  gene_recombination_probability = 0.9,
  mutation_probability = 0.8,
  gene_mutation_probability = 0.7,
  revert_change_probability = 0.3,
  max_generations = 100,
  population_size = 20,
  validity_threshold = function(x) quantile(x, 0.5, names = FALSE),
  tol_iter = 3,
  verbose = FALSE,
  seed = NULL
)
}
\arguments{
\item{explainer}{A \code{survex::survival_explainer} object - model preprocessed by the \code{survex::explain()} function.}

\item{new_observation}{A data frame with a single row containing the observation for which counterfactuals are to be generated.}

\item{target_prediction_band}{An object of class \code{target_prediction_band} containing the target prediction band for the counterfactuals.}

\item{weights}{A numeric vector with non-negative values and sum(weights) > 0. Weights for each time point. Default is \code{survex::survival_weights(explainer, p=0, q=0)}, which assigns equal weights to all time points.}

\item{times}{A numeric vector with time points for which the predictions are to be calculated.}

\item{background_data}{A data frame with the background data used to train the model. If not provided, the data from the explainer object is used.}

\item{target_change}{A numeric vector with the target change in the prediction expresses as survival distance. If provided, the counterfactuals will be generated to change the prediction by this amount.}

\item{fixed_variables_indices}{A vector of positive integers indicating the indices of variables that should remain fixed in the counterfactuals.}

\item{numerical_variables_indices}{A vector of positive integers indicating the indices of variables that are numerical. Variable types are inferred from the background data if not provided.}

\item{categorical_variables_indices}{A vector of positive integers indicating the indices of variables that are categorical. Variable types are inferred from the background data if not provided.}

\item{integer_variables_indices}{A vector of positive integers indicating the indices of variables that are integer. Variable types are inferred from the background data if not provided.}

\item{binary_variables_indices}{A vector of positive integers indicating the indices of variables that are binary. Variable types are inferred from the background data if not provided.}

\item{plausible_data_range}{A matrix with 2 rows and length(numerical_variables_indices) columns indicating the plausible range for each numerical variable. If not provided, the range is inferred from the background data.}

\item{plausible_categorical_values}{A list with length(categorical_variables_indices) elements indicating the plausible values for each categorical variable. If not provided, the values are inferred from the background data.}

\item{initialization_type}{A character string indicating the initialization type for the counterfactuals. Options are "ice" (default) and "background". If "ice" is selected, the counterfactuals are initialized based on the survival adaptation of Individual Conditional Expctation curves. If "background" is selected, the counterfactuals are initialized using the background data.}

\item{data_distance_p}{A numeric value indicating the p parameter for the Gower distance calculation. Default is 1.}

\item{plausibility_k_neighbors}{A numeric value indicating the number of neighbors for the plausibility calculation. Default is 5.}

\item{sbx_eta}{A numeric value indicating the eta parameter for the simulated binary crossover (SBX) operator. Default is 5.}

\item{gene_recombination_probability}{A numeric value indicating the probability of recombination for each gene. Default is 0.9.}

\item{mutation_probability}{A numeric value indicating the probability of mutation for each individual. Default is 0.8.}

\item{gene_mutation_probability}{A numeric value indicating the probability of mutation for each gene. Default is 0.7.}

\item{revert_change_probability}{A numeric value indicating the probability of reverting the change in each gene. Default is 0.3.}

\item{max_generations}{A numeric value indicating the maximum number of generations for the genetic algorithm. Default is 100.}

\item{population_size}{A numeric value indicating the population size for the genetic algorithm. Default is 20.}

\item{validity_threshold}{A function or a numeric value indicating the threshold for the validity of the counterfactuals. If validity value is greater than the threshold, the counterfactual candidates are considered invalid and additionally penalized. Default is the median of the validity values.}

\item{tol_iter}{A numeric value indicating the number of iterations with insufficient progress to stop the search. Default is 3.}

\item{verbose}{A logical value indicating whether to print the progress of the search. Default is FALSE.}

\item{seed}{An integer value to set the seed for reproducibility.}
}
\description{
SurvMOC is a time-dependent model-agnostic counterfactual explanation method for survival analysis models. It generates counterfactual examples using multi-objective optimization.
}
