% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCFex-package.R
\docType{package}
\name{survCFex-package}
\alias{survCFex}
\alias{survCFex-package}
\title{survCFex: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mateusz Krzyziński \email{krzyzinskimateusz23@gmail.com}

}
\keyword{internal}
