% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_distance.R
\name{survival_distance}
\alias{survival_distance}
\title{Calculate survival distance between a survival prediction and other survival predictions}
\usage{
survival_distance(
  sf,
  sf_others,
  times,
  weights = rep(1/length(times), length(times))
)
}
\arguments{
\item{sf}{Time-dependent survival predictions for which the distance is calculated.}

\item{sf_others}{Time-dependent survival predictions to which the distance is calculated.}

\item{times}{A numeric vector of time points at which the distance is calculated.}

\item{weights}{A numeric vector of weights assigned to each time point.}
}
\description{
Calculate survival distance between a survival prediction and other survival predictions
}
