% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{survival_weights}
\alias{survival_weights}
\title{Calculate Fleming-Harrington survival weights}
\usage{
survival_weights(
  explainer,
  times = explainer$times,
  p = 0,
  q = 0,
  add_mean = FALSE
)
}
\arguments{
\item{explainer}{A \code{survex::survival_explainer} object - model preprocessed by the \code{survex::explain()} function.}

\item{times}{A numeric vector of time points at which the weights are calculated.}

\item{p}{A numeric value being a weight used for the power of Kaplan-Meier estimator values.}

\item{q}{A numeric value being a weight used for the power of 1 - Kaplan-Meier estimator values.}

\item{add_mean}{A logical value indicating whether to add the mean of the weights to the weights.}
}
\description{
Calculate Fleming-Harrington survival weights
}
