% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate_cumulative_hazard_to_survival}
\alias{translate_cumulative_hazard_to_survival}
\title{Translate cumulative hazard function to survival function}
\usage{
translate_cumulative_hazard_to_survival(cumulative_hazard_function)
}
\arguments{
\item{cumulative_hazard_function}{A numeric vector of cumulative hazard values.}
}
\description{
Translate cumulative hazard function to survival function
}
