% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate_survival_to_cumulative_hazard}
\alias{translate_survival_to_cumulative_hazard}
\title{Translate survival function to cumulative hazard function}
\usage{
translate_survival_to_cumulative_hazard(survival_function)
}
\arguments{
\item{survival_function}{A numeric vector of survival probabilities.}
}
\description{
Translate survival function to cumulative hazard function
}
