% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_band.R
\name{translate_target_prediction_band}
\alias{translate_target_prediction_band}
\title{Translate target prediction band to the other output type}
\usage{
translate_target_prediction_band(target_prediction_band)
}
\arguments{
\item{target_prediction_band}{An object of class \code{target_prediction_band} containing the target prediction band.}
}
\description{
Translate target prediction band from survival to cumulative hazard or vice versa.
}
