% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovalev_method.R
\name{kovalev_method}
\alias{kovalev_method}
\title{Kovalev method for counterfactual explanation}
\usage{
kovalev_method(
  explainer,
  new_observation,
  r,
  num_iter = 200,
  num_particles = 1000,
  verbose = 10,
  C = 1e+06,
  w = 0.729,
  c1 = 1.4945,
  c2 = 1.4945,
  tol_iter = 10,
  seed = NULL
)
}
\arguments{
\item{explainer}{a survival explainer object}

\item{new_observation}{a new observation for which we want to find a counterfactual}

\item{r}{target difference between the mean time to survival of the counterfactual and the original observation}

\item{num_iter}{number of iterations in PSO algorithm}

\item{num_particles}{number of particles in PSO algorithm}

\item{verbose}{if integer, print loss every verbose iterations}

\item{C}{penalty parameter for the loss function (used to enforce the counterfactuality)}

\item{w}{inertia weight in PSO algorithm}

\item{c1}{cognitive parameter in PSO algorithm}

\item{c2}{social parameter in PSO algorithm}

\item{seed}{seed for random number generation}
}
\description{
This function implements themethod for counterfactual explanations of survival models proposed by Kovalev et al.
}
