% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCFex-package.R
\docType{package}
\name{survCFex-package}
\alias{survCFex}
\alias{survCFex-package}
\title{survCFex: Counterfactual Explanations for Survival Analysis Models}
\description{
The survCFex package offers counterfactual explanation methods for survival analysis models' predictions. It includes SurvMOC and ST-PT methods, model-agnostic and model-specific techniques, respectively. Both these methods ensure the time-dependent explainability.
}
\author{
\strong{Maintainer}: Mateusz Krzyziński \email{krzyzinskimateusz23@gmail.com}

Other contributors:
\itemize{
  \item Przemyslaw Biecek [thesis advisor]
}

}
\keyword{internal}
