% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCFex-package.R
\docType{package}
\name{survCFex-package}
\alias{survCFex}
\alias{survCFex-package}
\title{survCFex: Counterfactual Explanations for Survival Analysis Models}
\description{
The survCFex package offers counterfactual explanation methods for survival analysis. It includes SurvMOC, a model-agnostic approach, and ST-PT method, specifically designed for tree-based ensemble models. These methods generate plausible and actionable counterfactuals to explain predicted survival outcomes.
}
\author{
\strong{Maintainer}: Mateusz Krzyziński \email{krzyzinskimateusz23@gmail.com}

Other contributors:
\itemize{
  \item Przemyslaw Biecek [thesis advisor]
}

}
\keyword{internal}
